+++ castle/cms/static/components/tours/welcome.js
@@ -6,40 +6,40 @@ define([
   'use strict';
 
   var steps = [{
-      intro: "Welcome to Castle CMS. We've created a guide to help you get acquainted."
+      intro: "Welcome to CastleCMS. We've created a guide to help you get acquainted."
     }, {
       element: document.querySelector('.castle-toolbar-container-side'),
-      intro: 'The toolbar is where you will initiate all CMS changes.',
+      intro: 'The toolbar lets you make all content changes.',
       position: 'right'
     }, {
-      element: document.querySelector('.castle-toolbar-manage-content'),
-      intro: 'To view site structure, batch edit and organize content, use the Contents menu.',
+      element: document.querySelector('.castle-toolbar-folder_contents'),
+      intro: 'Use the Contents button to view your site \'s structure, organize content, and make batch edits.',
       position: 'right'
     }, {
       element: document.querySelector('.castle-toolbar-edit'),
-      intro: 'Click the edit button to edit the current page you are viewing',
+      intro: 'Click the Edit button to edit the current page or change its propertings and settings.',
       position: 'right',
       valid: function(){
         return $('.castle-toolbar-edit').length > 0;
       }
     }, {
       element: document.querySelector('.castle-toolbar-add'),
-      intro: 'To add a new page to the site, use the "Add new" menu item.',
+      intro: 'Use the Add button to add content, whether a page, folder, news item, or more.',
       position: 'right',
       valid: function(){
         return $('.castle-toolbar-add').length > 0;
       }
     }, {
       element: document.querySelector('.castle-toolbar-slots'),
-      intro: 'To add content above, below or on the sides of your content body, use the Slots menu item. ' +
-             "The amount of slots available depends on the site's theme.",
+      intro: 'Use the Slots button to add content above, below or on the sides of the main content area. ' +
+             "The number and location of slots available depends on the site's theme.",
       position: 'right',
       valid: function(){
         return $('.castle-toolbar-slots').length > 0;
       }
     }, {
       element: document.querySelector('.castle-btn-dropdown-user'),
-      intro: 'The user menu is for user specific functions like going back to your dashboard, ' +
+      intro: 'The user menu contains user-specific functions such as viewing your dashboard, ' +
              'changing preferences and logging out.',
       position: 'bottom'
     }];