+++ castle/cms/utils.py
@@ -489,6 +489,15 @@ def get_backend_url():
         backend_url = api.portal.get().absolute_url()
     return backend_url[0]
 
+def no_backend_url(value):
+    try:
+        url = str(value)
+        backend = get_backend_url()
+        if backend in url:
+            return False
+        return True
+    except:
+        return True
 
 def parse_query_from_data(data, context=None):
     if context is None: